requirements.txt
requests
numpy
psycopg2
scikit-learn
pillow
torch
torchvision
transformers
